Refactor: export function which is responsible for login to actions file
