+++ Front/clinicfront/src/reducers/info.js
@@ -1,7 +1,6 @@
 const info = (state = {uuid: "empty", role:"empty"} , action) => {
     switch (action.type) {
         case 'SET_USER_DETAILS':
-            console.log(action.userDetails.uuid);
             return {userDetails: {
                 uuid: action.userDetails.uuid,
                 role: action.userDetails.role
